feat: add ConverterHistoryList component and integrate it into NotebookHistoryModal and SaveDatasetModal
