+++ DashAI/front/src/components/notebooks/utils.js
@@ -1,0 +1,5 @@
+export const formatScope = (scope) => {
+  const cols = scope.columns?.length ? scope.columns.join(", ") : "All";
+  const rows = scope.rows?.length ? scope.rows.join(", ") : "All";
+  return `Columns: ${cols} | Rows: ${rows}`;
+};
